feat: add spinner for loading wasm
