+++ src/utils/setTimeOutPromise.ts
@@ -1,0 +1,5 @@
+export const setTimeOutPromise = (ms: number): Promise<void> => {
+  return new Promise((resolve) => {
+    setTimeout(resolve, ms);
+  });
+};
